Offload dynamic variables to shared_variables. Allow to regenerate variables using load_methods_and_modules, init, and load_profiles methods.
